+++ java/client/src/main/java/org/signal/libsignal/net/Network.java
@@ -8,8 +8,11 @@ package org.signal.libsignal.net;
 import static org.signal.libsignal.internal.FilterExceptions.filterExceptions;
 
 import java.io.IOException;
+import java.util.Collections;
+import java.util.Map;
 import java.util.concurrent.ExecutionException;
 import java.util.function.Consumer;
+import org.signal.libsignal.internal.BridgedStringMap;
 import org.signal.libsignal.internal.CompletableFuture;
 import org.signal.libsignal.internal.Native;
 import org.signal.libsignal.internal.NativeHandleGuard;
@@ -39,8 +42,12 @@ public class Network {
   private final ConnectionManager connectionManager;
 
   public Network(Environment env, String userAgent) {
+    this(env, userAgent, Collections.emptyMap());
+  }
+
+  public Network(Environment env, String userAgent, Map<String, String> remoteConfig) {
     this.tokioAsyncContext = new TokioAsyncContext();
-    this.connectionManager = new ConnectionManager(env, userAgent);
+    this.connectionManager = new ConnectionManager(env, userAgent, remoteConfig);
   }
 
   /**
@@ -129,6 +136,11 @@ public class Network {
    */
   public void setCensorshipCircumventionEnabled(boolean enabled) {
     this.connectionManager.setCensorshipCircumventionEnabled(enabled);
+  }
+
+  /** Updates the remote config settings used by libsignal. */
+  public void setRemoteConfig(Map<String, String> remoteConfig) {
+    this.connectionManager.setRemoteConfig(remoteConfig);
   }
 
   /**
@@ -242,8 +254,10 @@ public class Network {
   static class ConnectionManager extends NativeHandleGuard.SimpleOwner {
     private final Environment environment;
 
-    private ConnectionManager(Environment env, String userAgent) {
-      super(Native.ConnectionManager_new(env.value, userAgent));
+    private ConnectionManager(Environment env, String userAgent, Map<String, String> remoteConfig) {
+      super(
+          new BridgedStringMap(remoteConfig)
+              .guardedMap(map -> Native.ConnectionManager_new(env.value, userAgent, map)));
       this.environment = env;
     }
 
@@ -290,6 +304,12 @@ public class Network {
 
     private void setCensorshipCircumventionEnabled(boolean enabled) {
       guardedRun(h -> Native.ConnectionManager_set_censorship_circumvention_enabled(h, enabled));
+    }
+
+    private void setRemoteConfig(Map<String, String> remoteConfig) {
+      new BridgedStringMap(remoteConfig)
+          .guardedRun(
+              map -> this.guardedRun(h -> Native.ConnectionManager_set_remote_config(h, map)));
     }
 
     @Override