+++ swift/Sources/LibSignalClient/Net.swift
@@ -26,9 +26,9 @@ public class Net {
     public static let signalTlsProxyScheme = "org.signal.tls"
 
     /// Creates a new `Net` instance that enables interacting with services in the given Signal environment.
-    public init(env: Environment, userAgent: String) {
+    public init(env: Environment, userAgent: String, remoteConfig: [String: String] = [:]) {
         self.asyncContext = TokioAsyncContext()
-        self.connectionManager = ConnectionManager(env: env, userAgent: userAgent)
+        self.connectionManager = ConnectionManager(env: env, userAgent: userAgent, remoteConfig: remoteConfig)
     }
 
     /// Sets the proxy host to be used for all new connections (until overridden).
@@ -100,6 +100,11 @@ public class Net {
     /// CC is off by default.
     public func setCensorshipCircumventionEnabled(_ enabled: Bool) {
         self.connectionManager.setCensorshipCircumventionEnabled(enabled)
+    }
+
+    /// Updates the remote config settings used by libsignal.
+    public func setRemoteConfig(_ remoteConfig: [String: String]) {
+        self.connectionManager.setRemoteConfig(remoteConfig)
     }
 
     /// Notifies libsignal that the network has changed.
@@ -294,9 +299,11 @@ internal class ConnectionManager: NativeHandleOwner<SignalMutPointerConnectionMa
         }
     }
 
-    convenience init(env: Net.Environment, userAgent: String) {
+    convenience init(env: Net.Environment, userAgent: String, remoteConfig: [String: String]) {
         var handle = SignalMutPointerConnectionManager()
-        failOnError(signal_connection_manager_new(&handle, env.rawValue, userAgent))
+        remoteConfig.withBridgedStringMap { remoteConfig in
+            failOnError(signal_connection_manager_new(&handle, env.rawValue, userAgent, remoteConfig))
+        }
         self.init(owned: NonNull(handle)!)
     }
 
@@ -340,6 +347,14 @@ internal class ConnectionManager: NativeHandleOwner<SignalMutPointerConnectionMa
     internal func setCensorshipCircumventionEnabled(_ enabled: Bool) {
         self.withNativeHandle {
             failOnError(signal_connection_manager_set_censorship_circumvention_enabled($0.const(), enabled))
+        }
+    }
+
+    internal func setRemoteConfig(_ remoteConfig: [String: String]) {
+        remoteConfig.withBridgedStringMap { remoteConfig in
+            self.withNativeHandle {
+                failOnError(signal_connection_manager_set_remote_config($0.const(), remoteConfig))
+            }
         }
     }