+++ RELEASE_NOTES.md
@@ -1,3 +1,5 @@
 v0.69.1
 
 - Node (GSE): Implement toToken() and encryptUserId() for CallLinkSecretParams
+
+- The Net class (Network in Java) now stores a string-map of "remote config" information, intended for the same sort of server-provided configuration that the apps already have.